+++ awslimitchecker/services/ec2.py
@@ -47,6 +47,8 @@ from ..limit import AwsLimit
 
 logger = logging.getLogger(__name__)
 
+RI_NO_AZ = 'xxREGIONAL_BENEFIT-NO_AZxx'
+
 
 class _Ec2Service(_AwsService):
 
@@ -78,6 +80,7 @@ class _Ec2Service(_AwsService):
         # update our limits with usage
         inst_usage = self._instance_usage()
         res_usage = self._get_reserved_instance_count()
+        logger.debug('Reserved instance count: %s', res_usage)
         # subtract reservations from instance usage
         ondemand_usage = defaultdict(int)
         for az in inst_usage:
@@ -96,6 +99,20 @@ class _Ec2Service(_AwsService):
                     # we have unused reservations
                     continue
                 ondemand_usage[i_type] += od
+        # subtract any "Regional Benefit" AZ-less reservations
+        for i_type in ondemand_usage.keys():
+            if RI_NO_AZ in res_usage and i_type in res_usage[RI_NO_AZ]:
+                logger.debug('Subtracting %d AZ-less "Regional Benefit" '
+                             'Reserved Instances from %d running %s instances',
+                             res_usage[RI_NO_AZ][i_type],
+                             ondemand_usage[i_type], i_type)
+                # we have Regional Benefit reservations for this type;
+                # we don't want to show negative usage, even if we're
+                #  over-reserved
+                ondemand_usage[i_type] = max(
+                    ondemand_usage[i_type] - res_usage[RI_NO_AZ][i_type],
+                    0
+                )
         total_instances = 0
         for i_type, usage in ondemand_usage.items():
             key = 'Running On-Demand {t} instances'.format(
@@ -200,6 +217,9 @@ class _Ec2Service(_AwsService):
                 logger.debug("Skipping ReservedInstance %s with state %s",
                              x['ReservedInstancesId'], x['State'])
                 continue
+            if 'AvailabilityZone' not in x:
+                # "Regional Benefit" AZ-less reservation
+                x['AvailabilityZone'] = RI_NO_AZ
             if x['AvailabilityZone'] not in az_to_res:
                 az_to_res[x['AvailabilityZone']] = deepcopy(reservations)
             az_to_res[x['AvailabilityZone']][
@@ -299,12 +319,18 @@ class _Ec2Service(_AwsService):
         # (On-Demand, Reserved, Spot)
         default_limits = (20, 20, 5)
         special_limits = {
+            'm4.4xlarge': (10, 20, 5),
+            'm4.10xlarge': (5, 20, 5),
+            'm4.16xlarge': (5, 20, 5),
             'c4.4xlarge': (10, 20, 5),
             'c4.8xlarge': (5, 20, 5),
             'cg1.4xlarge': (2, 20, 5),
             'hi1.4xlarge': (2, 20, 5),
             'hs1.8xlarge': (2, 20, 0),
             'cr1.8xlarge': (2, 20, 5),
+            'p2.xlarge': (1, 20, 5),
+            'p2.8xlarge': (1, 20, 5),
+            'p2.16xlarge': (1, 20, 5),
             'g2.2xlarge': (5, 20, 5),
             'g2.8xlarge': (2, 20, 5),
             'r3.4xlarge': (10, 20, 5),
@@ -315,8 +341,6 @@ class _Ec2Service(_AwsService):
             'i2.8xlarge': (2, 20, 0),
             'd2.4xlarge': (10, 20, 5),
             'd2.8xlarge': (5, 20, 5),
-            'm4.4xlarge': (10, 20, 5),
-            'm4.10xlarge': (5, 20, 5)
         }
         limits = {}
         for i_type in self._instance_types():
@@ -559,7 +583,8 @@ class _Ec2Service(_AwsService):
             'm4.xlarge',
             'm4.2xlarge',
             'm4.4xlarge',
-            'm4.10xlarge'
+            'm4.10xlarge',
+            'm4.16xlarge'
         ]
 
         PREV_GENERAL_TYPES = [
@@ -576,6 +601,8 @@ class _Ec2Service(_AwsService):
             'r3.2xlarge',
             'r3.4xlarge',
             'r3.8xlarge',
+            'x1.16xlarge',
+            'x1.32xlarge'
         ]
 
         PREV_MEMORY_TYPES = [
@@ -602,6 +629,12 @@ class _Ec2Service(_AwsService):
             'c1.medium',
             'c1.xlarge',
             'cc2.8xlarge',
+        ]
+
+        ACCELERATED_COMPUTE_TYPES = [
+            'p2.xlarge',
+            'p2.8xlarge',
+            'p2.16xlarge'
         ]
 
         STORAGE_TYPES = [
@@ -639,6 +672,7 @@ class _Ec2Service(_AwsService):
             PREV_MEMORY_TYPES +
             COMPUTE_TYPES +
             PREV_COMPUTE_TYPES +
+            ACCELERATED_COMPUTE_TYPES +
             STORAGE_TYPES +
             PREV_STORAGE_TYPES +
             DENSE_STORAGE_TYPES +