+++ awslimitchecker/tests/support.py
@@ -146,6 +146,10 @@ class LogRecordHelper(object):
                     r.funcName == '_find_usage_nat_gateways' and
                     'perhaps NAT service does not exist in this regi' in r.msg):
                 continue
+            if (r.levelno == logging.ERROR and r.module == 'firehose' and
+                    r.funcName == 'find_usage' and 'perhaps the Firehose '
+                    'service is not available in this region' in r.msg):
+                continue
             res.append('%s:%s.%s (%s:%s) %s - %s %s' % (
                 r.name,
                 r.module,