+++ awslimitchecker/versioncheck.py
@@ -46,7 +46,7 @@ import locale
 
 if sys.version_info >= (3, 3):
     from subprocess import DEVNULL
-else:
+else:  # unreachable under 3.4+ - pragma: no cover
     DEVNULL = open(os.devnull, 'wb')
 
 try:
@@ -273,7 +273,7 @@ def _check_output(args, stderr=None):
         if p.returncode != 0:
             raise subprocess.CalledProcessError(p.returncode, args)
     else:
-        res = subprocess.check_output(args, stderr=stderr)
+        res = subprocess.check_output(args, stderr=stderr)  # pragma: no cover
         if sys.version_info >= (3, 0):
             res = res.decode(locale.getdefaultlocale()[1])
     return res