+++ CHANGES.rst
@@ -1,6 +1,45 @@
 Changelog
 =========
 
+0.6.0 (2016-11-12)
+------------------
+
+This release has a breaking change. The ``VPC`` ``NAT gateways`` has been renamed
+to ``NAT Gateways per AZ`` and its ``get_current_usage()`` method will now return
+a list with multiple items. See the changelog entry for #214 below.
+
+This release requires the following new IAM permissions to function:
+
+* ``firehose:ListDeliveryStreams``
+
+* `#217 <https://github.com/jantman/awslimitchecker/issues/217>`_ - add support
+  for new/missing EC2 instance types: ``m4.16xlarge``, ``x1.16xlarge``, ``x1.32xlarge``,
+  ``p2.xlarge``, ``p2.8xlarge``, ``p2.16xlarge``.
+* `#215 <https://github.com/jantman/awslimitchecker/issues/215>`_ - support
+  "Regional Benefit" Reserved Instances that have no specific AZ set on them. Per
+  AWS, these are exempt from On-Demand Running Instances limits like all other
+  RIs.
+* `#214 <https://github.com/jantman/awslimitchecker/issues/214>`_ - The VPC "NAT gateways"
+  limit incorrectly calculated usage for the entire region, while the limit is
+  actually per-AZ. It also had strange capitalization that confused users. The name
+  has been changed to "NAT Gateways per AZ" and the usage is now correctly calculated
+  per-AZ instead of region-wide.
+* `#221 <https://github.com/jantman/awslimitchecker/issues/221>`_ /
+  `PR #222 <https://github.com/jantman/awslimitchecker/pull/222>`_ - Fix bug
+  in handling of STS Credentials where they are cached permanently in
+  ``connectable.Connectable.credentials``, and new AwsLimitChecker instances
+  in the same Python process reuse the first set of STS credentials. This is
+  fixed by storing the Account ID as part of
+  ``connectable.ConnectableCredentials`` and getting new STS creds if the cached
+  account ID does not match the current ``account_id`` on the ``Connectable``
+  object.
+* `PR #216 <https://github.com/jantman/awslimitchecker/pull/216>`_ - add new
+  "Firehose" service with support for "Delivery streams per region" limit.
+* `#213 <https://github.com/jantman/awslimitchecker/issues/213>`_ /
+  `PR #188 <https://github.com/jantman/awslimitchecker/pull/188>`_ - support
+  AWS cross-sdk credential file profiles via ``-P`` / ``--profile``, like
+  awscli.
+
 0.5.1 (2016-09-25)
 ------------------